requirements.txt
torch>=1.9.0
torchvision>=0.10.0
timm>=0.5.4
numpy>=1.20.0
opencv-python>=4.5.0
scikit-learn>=0.24.0
tqdm>=4.60.0
Pillow>=8.3.0